# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/advertisers/{advertiserId}/campaigns/{campaignId}/ad-groups/{adGroupId}" : {
      "get" : {
        "tags" : [ "ad-group-controller" ],
        "operationId" : "getAdGroup",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "adGroupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAdGroupResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "ad-group-controller" ],
        "operationId" : "updateAdGroup",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "adGroupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdGroupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAdGroupResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "ad-group-controller" ],
        "operationId" : "deleteAdGroup",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "adGroupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/creatives/{creativeId}/audit" : {
      "post" : {
        "tags" : [ "ad-creative-controller" ],
        "operationId" : "auditCreative",
        "parameters" : [ {
          "name" : "creativeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "object"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/advertisers/{advertiserId}/campaigns/{campaignId}/ad-groups" : {
      "get" : {
        "tags" : [ "ad-group-controller" ],
        "operationId" : "listAdGroups",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageAdGroupResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "ad-group-controller" ],
        "operationId" : "createAdGroup",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdGroupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAdGroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advertisers/{advertiserId}/campaigns/{campaignId}/ad-groups/{adGroupId}/creatives" : {
      "get" : {
        "tags" : [ "ad-creative-controller" ],
        "operationId" : "listCreatives",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "adGroupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAdCreativeResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "ad-creative-controller" ],
        "operationId" : "createCreative",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "adGroupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdCreativeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAdCreativeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advertisers/{advertiserId}/campaigns/{campaignId}/ad-groups/{adGroupId}/creatives/upload-url" : {
      "post" : {
        "tags" : [ "ad-creative-controller" ],
        "operationId" : "generateUploadUrl",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "adGroupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "fileName",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "contentType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AdGroupRequest" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "geoTargeting" : {
            "$ref" : "#/components/schemas/GeoTargeting"
          },
          "daypartingConfig" : {
            "$ref" : "#/components/schemas/DaypartingConfig"
          },
          "contextualKeywords" : {
            "$ref" : "#/components/schemas/ContextualKeywords"
          },
          "brandSafetyBlocklist" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
      "ContextualKeywords" : {
        "required" : [ "keywords" ],
        "type" : "object",
        "properties" : {
          "keywords" : {
            "maxItems" : 400,
            "minItems" : 0,
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "Daypart" : {
        "required" : [ "dayOfWeek", "endTime", "startTime" ],
        "type" : "object",
        "properties" : {
          "dayOfWeek" : {
            "type" : "string"
          },
          "startTime" : {
            "type" : "string"
          },
          "endTime" : {
            "type" : "string"
          }
        }
      },
      "DaypartingConfig" : {
        "required" : [ "dayparts" ],
        "type" : "object",
        "properties" : {
          "dayparts" : {
            "maxItems" : 400,
            "minItems" : 0,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Daypart"
            }
          }
        }
      },
      "GeoTargeting" : {
        "required" : [ "regions" ],
        "type" : "object",
        "properties" : {
          "regions" : {
            "maxItems" : 400,
            "minItems" : 0,
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "AdGroupResponse" : {
        "required" : [ "active", "campaignId", "createdAt", "id", "name", "updatedAt" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "campaignId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "geoTargeting" : {
            "$ref" : "#/components/schemas/GeoTargeting"
          },
          "daypartingConfig" : {
            "$ref" : "#/components/schemas/DaypartingConfig"
          },
          "contextualKeywords" : {
            "$ref" : "#/components/schemas/ContextualKeywords"
          },
          "brandSafetyBlocklist" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "active" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseAdGroupResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdGroupResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseVoid" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "AdCreativeRequest" : {
        "required" : [ "format" ],
        "type" : "object",
        "properties" : {
          "format" : {
            "type" : "string",
            "enum" : [ "BANNER", "CAROUSEL", "VIDEO", "VIDEO_VAST", "NATIVE" ]
          },
          "assetUrl" : {
            "type" : "string"
          },
          "vastXml" : {
            "type" : "string"
          }
        }
      },
      "AdCreativeResponse" : {
        "required" : [ "adGroupId", "assetUrl", "auditStatus", "createdAt", "format", "id", "updatedAt" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "adGroupId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "format" : {
            "type" : "string",
            "enum" : [ "BANNER", "CAROUSEL", "VIDEO", "VIDEO_VAST", "NATIVE" ]
          },
          "assetUrl" : {
            "type" : "string"
          },
          "vastXml" : {
            "type" : "string"
          },
          "auditStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "REJECTED" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseAdCreativeResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdCreativeResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseString" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          }
        }
      },
...
      "PageableObject" : {
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "unpaged" ],
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "unpaged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "SortObject" : {
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponsePageAdGroupResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageAdGroupResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageAdGroupResponse" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AdGroupResponse"
            }
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "ApiResponseListAdCreativeResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AdCreativeResponse"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      }
    }
  }
}